Plot dark and light themed mulit armed bandits
